Move XMPP login prompt handling to AuthHandler
